refactor(revision): rename files for clarity
manager.go → helpers.go (stateless CR builders and error checks)
revision_manager.go → manager.go (Manager implementation)
